requirements.txt
aiofiles==0.8.0
aiomysql==0.1.1
asyncio-redis==0.16.0
decorator==5.1.1
elasticsearch6==6.8.2
Faker==14.1.0
greenlet==1.1.3
httptools==0.4.0
importlib-metadata==4.12.0
multidict==6.0.2
pbr==5.10.0
PyJWT==2.4.0
PyMySQL==1.0.2
PyNaCl==1.5.0
python-dateutil==2.8.2
sanic==22.6.2
sanic-routing==22.3.0
sanic-session==0.8.0
six==1.16.0
SQLAlchemy==1.4.40
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
Tempita==0.5.2
typing_extensions==4.3.0
ujson==5.4.0
urllib3==1.26.12
websockets==10.3
zipp==3.8.1
